Move `ReplicationMessage` into a separate module
